* update Trinity-RFT train executor to an specific dir
